Refactor Order classes to use BigDecimal for price representation; update related tests and OrderFactory for consistency

+++ src/test/java/dev/rahatali/orderbook/entities/OrderBookTest.java
@@ -4,6 +4,8 @@ import dev.rahatali.orderbook.enums.Strategy;
 import dev.rahatali.orderbook.enums.Type;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -16,11 +18,11 @@ class OrderBookTest {
         assertTrue(orderBook.isEmpty(null));
 
         // Test with order book with bids
-        orderBook.add(new Bid(1, 100f, 10, Strategy.LIMIT));
+        orderBook.add(new Bid(1, BigDecimal.valueOf(100), 10, Strategy.LIMIT));
         assertFalse(orderBook.isEmpty(Type.BID));
 
         // Test with order book with asks
-        orderBook.add(new Ask(1, 100f, 10, Strategy.LIMIT));
+        orderBook.add(new Ask(1, BigDecimal.valueOf(100), 10, Strategy.LIMIT));
         assertFalse(orderBook.isEmpty(Type.ASK));
 
         // Test with order book with bids and asks
@@ -30,8 +32,8 @@ class OrderBookTest {
     @Test
     void testMatchWithTwoOrders() {
         OrderBook orderBook = new OrderBook();
-        orderBook.add(new Bid(1, 100f, 10, Strategy.LIMIT));
-        orderBook.add(new Ask(2, 100f, 10, Strategy.LIMIT));
+        orderBook.add(new Bid(1, BigDecimal.valueOf(100), 10, Strategy.LIMIT));
+        orderBook.add(new Ask(2, BigDecimal.valueOf(100), 10, Strategy.LIMIT));
         assertTrue(orderBook.match());
         System.out.println(orderBook);
         assertTrue(orderBook.isEmpty(null));
